Split models into unsee and Alertmanager specific files
